Adjusted several functions
Added the workflow functions for after clicking connect: user sees a div with graph options and 'add graph button' and an invisible 'graph div' is created above this displayed graph div. Upon clicking, appends a new div with graph information to the invisible graph. New div includes a delete button. Upon clicking, triggers same callback as the 'add_graph' callback but with a check decides whether to add a graph or to delete the div of the corresponding dynamic index of the graph the user wants to delete.